webapp/requirements.txt
flask==2.0.1
psycopg2-binary==2.9.1
toml==0.10.2
gunicorn==20.1.0
